+++ src/app/features/mapping-regions-to-chief-mentor/regioncreation/regioncreation.component.ts
@@ -1,11 +1,13 @@
 import { SelectionModel } from '@angular/cdk/collections';
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { MatDialog } from '@angular/material/dialog';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { map, Observable, startWith } from 'rxjs';
+import { DialogPopupComponent } from 'src/app/shared/dialog-popup/dialog-popup.component';
 import { MappingRegionsToChiefMentorService } from '../services/mapping-regions-to-chief-mentor.service';
 
 @Component({
@@ -25,9 +27,11 @@ export class RegioncreationComponent implements OnInit {
   displaycontent: boolean = false
 
   iseditable: boolean = false
+  issubmit: boolean = true
 
 
 
+  pageno: number = 1
 
   filterData: any;
   gridData = [];
@@ -45,7 +49,8 @@ export class RegioncreationComponent implements OnInit {
     private service: MappingRegionsToChiefMentorService,
     private _snackbar: MatSnackBar,
     private formbuilder: FormBuilder,
-    private _snackBar: MatSnackBar
+    private _snackBar: MatSnackBar,
+    private dialog: MatDialog,
   ) {
 
     this.regionfilterform = this.formbuilder.group({
@@ -66,6 +71,17 @@ export class RegioncreationComponent implements OnInit {
       paginator: this.paginator,
       sort: this.sort
     }
+    this.service.getcountry().subscribe({
+      next: (response) => {
+
+        this.countryList = response
+        this.countryList = this.countryList.map((ele: any) => ele.countryName)
+      },
+      error: (error) => {
+        console.error(error.message);
+
+      }
+    })
   }
 
   ngOnInit(): void {
@@ -74,11 +90,11 @@ export class RegioncreationComponent implements OnInit {
 
 
   getregiondata() {
-    this.service.getregion().subscribe({
+    this.service.getallregion().subscribe({
       next: (value) => {
         this.data = value
+
         this.data = this.data.reverse()
-        this.countryList = [...new Set(this.data.map((ele: any) => ele.countryName))]
         this.statelist = [...new Set(this.data.map((ele: any) => ele.states))]
 
         const key = "partId"
@@ -96,16 +112,6 @@ export class RegioncreationComponent implements OnInit {
           startWith(''),
           map(value => this.statelist.filter((ele: any) => ele.toLowerCase().includes(value.state?.toLowerCase()))),
         )
-
-        const regionlist = [...new Set(this.data.map((ele: any) => ele.regionName))]
-
-
-
-        this.regionfilter = this.regionfilterform.valueChanges.pipe(
-          startWith(''),
-          map(value => regionlist.filter((ele: any) => ele.toLowerCase().includes(value.region?.toLowerCase()))),
-        )
-
         this.dataSource = new MatTableDataSource<any>(this.data)
         this.filterData.gridData = this.data;
         this.filterData.dataSource = this.dataSource;
@@ -152,27 +158,42 @@ export class RegioncreationComponent implements OnInit {
     )
   }
 
-  opensnackBar(data: any) {
-    this._snackbar.open(data.message, "Close")
+  openSnackBar(data: any) {
+    this._snackBar.open(data.message, 'Close', {
+      duration: 2 * 1000,
+    });
   }
 
   updatePagination(col: any) {
-    this.ngAfterViewInit()
+    this.filterData.dataSource.paginator = this.paginator;
+    this.filterData.dataSource.sort = this.sort;
   }
   compareselect(obj1: any, obj2: any) {
     return obj1 && obj2 && obj1 === obj2
   }
 
-  ngAfterViewInit() {
-    this.filterData.dataSource.paginator = this.paginator;
-    this.filterData.dataSource.sort = this.sort;
-
+  onpaginatechange(event: any) {
+    if (event.pageIndex === 0) {
+      this.pageno = 1
+      return
+    }
+    this.pageno = (event.pageIndex * event.pageSize) + 1
+    return
   }
   addregion() {
     this.displaycontent = !this.displaycontent
   }
   viewdetails(element: any) {
+    this.regionfilterform.setValue({
+      regionId: element.regionId,
+      region: element.regionName,
+      country: element.countryName,
+      part: element.partId.partId,
+      state: element.states
 
+    });
+    this.displaycontent = true
+    this.issubmit = false
   }
   deletedetails(id: any) {
 
@@ -180,16 +201,35 @@ export class RegioncreationComponent implements OnInit {
       "regionId": id,
     }
 
-    this.service.deleteregion(body).subscribe({
-      next: (response) => {
-        this.opensnackBar(response)
-        this.regionfilterform.reset()
-        this.getregiondata()
+
+    const dialogref = this.dialog.open(DialogPopupComponent, {
+      data: {
+        title: "Delete Confirmation",
+        message: "Are You Sure You Want To Delete this region ?"
       },
-      error: (error) => {
-        console.error(error.message);
-      }
+      width: "30%"
     })
+
+    dialogref.afterClosed().subscribe(data => {
+      if (data) {
+        this.service.deleteregion(body).subscribe({
+          next: (response) => {
+            this.openSnackBar(response)
+            this.regionfilterform.reset()
+            this.getregiondata()
+          },
+          error: (error) => {
+            console.error(error.message);
+          }
+        })
+        return
+      }
+
+    })
+
+
+
+    return
 
   }
   editdetails(element: any) {
@@ -204,12 +244,15 @@ export class RegioncreationComponent implements OnInit {
     });
     this.iseditable = true
     this.displaycontent = true
+    this.issubmit = true
   }
 
   reseteditable() {
     this.regionfilterform.reset()
     this.iseditable = false
     this.displaycontent = !this.displaycontent
+    this.issubmit = true
+
   }
 
 
@@ -232,21 +275,23 @@ export class RegioncreationComponent implements OnInit {
     }
 
     if (this.iseditable) {
-      //editable
 
+      const partname = this.data.filter((ele: any) => ele.partId.partId === part)[0].partId.partName
       const body = {
         "regionId": regionId,
         "regionName": region,
         "countryName": country,
         "partId": {
-          "partId": part
+          "partId": part,
+          "partName": partname,
         },
         "states": state
       }
 
+
       this.service.updateregion(body).subscribe({
         next: (response) => {
-          this.opensnackBar(response)
+          this.openSnackBar(response)
           this.regionfilterform.reset()
           this.getregiondata()
         },
@@ -255,11 +300,12 @@ export class RegioncreationComponent implements OnInit {
 
         }
       })
+      return
     }
 
     this.service.postregion(body).subscribe({
       next: (response) => {
-        this.opensnackBar(response)
+        this.openSnackBar(response)
         this.regionfilterform.reset()
         this.getregiondata()
       },
